change date text field to label in ExecuteTaskViewController modify page control in RecordContentViewController

+++ Tastopia/Model/FirestoreManager.swift
@@ -131,7 +131,7 @@ struct WritingData: Codable {
     let number: Int
     let uid: String
     let userName: String
-    let date: Int
+    let date: Double
     let composition: String
     let images: [String]
     let agree: Int
